+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0-dev.1
+
+* Add support for specifying payment intent and user action in `BraintreePayPalRequest` (thank you to [nabinadhikari](https://github.com/nabinadhikari)!)
+* Clean up naming conventions and documentation
+
 ## 2.3.1
 
 * Update iOS dependencies (thank you to [jorgefspereira](https://github.com/jorgefspereira)!)